Separate the cluster related pacemaker-crowbar proxies
